requirements.txt
discord.py
requests
python-dotenv
humanize